Rename the total_bandwidth file for clarity
