+++ src/display/components/header_details.rs
