+++ src/display/components/mod.rs
@@ -1,11 +1,11 @@
 mod display_bandwidth;
+mod header_details;
 mod help_text;
 mod layout;
 mod table;
-mod total_bandwidth;
 
 pub use display_bandwidth::*;
+pub use header_details::*;
 pub use help_text::*;
 pub use layout::*;
 pub use table::*;
-pub use total_bandwidth::*;
